requirements.txt
beautifulsoup4>=4.12.0
langchain-core>=0.1.0
langchain-ollama>=0.0.1
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
